Add IPC for toggling window visiblity:
Add socket listeners to toggling window visiblity. This is to make up
for the lack of global support on wayland based compositors like
hyprland as registering global shortcuts doesn't work and in case of
hyprland bindings are manually. The commit exposed "toggle", "show" and
"hide" events through socket communcation for controlling the window
visibility.